requirements.txt
pandas
numpy==2.0.0
scikit-learn
matplotlib
seaborn
streamlit==1.38.0
google-generativeai
streamlit-authenticator
openpyxl
python-dotenv
streamlit-lottie
xgboost
ydata-profiling
langchain
langchain_community
langchain-google-genai